requirements.txt
jax==0.4.11
jaxlib==0.4.11
jax-metal==0.0.4
cupy-cuda12x==13.0.0
fastrlock==0.8.2
numpy==1.26.4
setuptools==69.2.0
tqdm==4.65.0
requests==2.31.0
